Use exact threshold for songs too short for scrobbling
For durations above 29963ms, the internal last.fm plugin seems to round
up and send the track to last.fm as thirty seconds. Or maybe that's
what last.fm accepts as "long enough". Either way, this should prevent
double scrobbles for the 37ms span between that and the previously set
value of exactly 30 seconds.

fixes #1